+++ prettier.config.mjs
@@ -1,0 +1,7 @@
+import prettierConfig from '@iobroker/eslint-config/prettier.config.mjs';
+
+export default {
+    ...prettierConfig,
+    singleQuote: true,
+};
+