+++ test/integration.js
@@ -1,0 +1,5 @@
+const path = require('path');
+const { tests } = require('@iobroker/testing');
+
+tests.integration(path.join(__dirname, '..'));
+